.: Moved Fts into their own individual files (Chapter 12)
