+++ changeLog.txt
@@ -4,6 +4,11 @@
 
 develop
 -------------------------------------------------------------
+- 2018-01-11                                                -
+-------------------------------------------------------------
+.: Moved Fts into their own individual files (Chapter 12)
+
+-------------------------------------------------------------
 - 2017-12-27                                                -
 -------------------------------------------------------------
 .: Add a fabfile for automated deploys (Chapter 11)
