+++ functional_tests/test_list_item_validation.py
@@ -1,0 +1,22 @@
+from selenium.webdriver.common.keys import Keys
+from unittest import skip
+from .base import FunctionalTest
+
+
+class ItemValidationTest(FunctionalTest):
+
+    def test_cannot_add_empty_list_items(self):
+        # Michelle goes to the home page and accidentally tries to submit
+        # an empty list item. She hits Enter on the empty iput box
+
+        # The home page refreshes, and there is an error message saying
+        # that list items cannot be blank
+
+        # She tries again with some text for the item, which now works
+
+        # Perversely, she now decides to submit a second blank list item,
+
+        # She receives a similar warning on the list page
+
+        # And she can correct it by filling some text in
+        self.fail('write me!')
